Added Roster Entry Engine with Tests

+++ src/NbaStats.Data/DataObjects/Transaction.cs
@@ -7,7 +7,7 @@ namespace NbaStats.Data.DataObjects
     public class Transaction
     {
         public long Id { get; set; }
-        public int PlayerId { get; set; }
+        public long PlayerId { get; set; }
         public int OldTeamId { get; set; }
         public int NewTeamId { get; set; }
     }
